Refactor plpc2 -> plpc
