+++ src/main/java/ch/ethz/matsim/av/dispatcher/single_fifo/SingleFIFODispatcher.java
@@ -7,7 +7,7 @@ import ch.ethz.matsim.av.dispatcher.AVVehicleAssignmentEvent;
 import ch.ethz.matsim.av.dispatcher.utils.SingleRideAppender;
 import ch.ethz.matsim.av.framework.AVModule;
 import ch.ethz.matsim.av.passenger.AVRequest;
-import ch.ethz.matsim.av.plcpc2.ParallelLeastCostPathCalculator;
+import ch.ethz.matsim.av.plcpc.ParallelLeastCostPathCalculator;
 import ch.ethz.matsim.av.schedule.AVTask;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
