+++ src/main/java/ch/ethz/matsim/av/framework/AVModule.java
@@ -10,9 +10,9 @@ import ch.ethz.matsim.av.dispatcher.single_fifo.SingleFIFODispatcher;
 import ch.ethz.matsim.av.dispatcher.single_heuristic.SingleHeuristicDispatcher;
 import ch.ethz.matsim.av.generator.AVGenerator;
 import ch.ethz.matsim.av.generator.PopulationDensityGenerator;
-import ch.ethz.matsim.av.plcpc2.DefaultParallelLeastCostPathCalculator;
-import ch.ethz.matsim.av.plcpc2.ParallelLeastCostPathCalculator;
-import ch.ethz.matsim.av.plcpc2.SerialLeastCostPathCalculator;
+import ch.ethz.matsim.av.plcpc.DefaultParallelLeastCostPathCalculator;
+import ch.ethz.matsim.av.plcpc.ParallelLeastCostPathCalculator;
+import ch.ethz.matsim.av.plcpc.SerialLeastCostPathCalculator;
 import ch.ethz.matsim.av.replanning.AVOperatorChoiceStrategy;
 import ch.ethz.matsim.av.routing.AVRoute;
 import ch.ethz.matsim.av.routing.AVRouteFactory;
